Fix multirank forward handling
In case of multiple ranks, the forwarding delay was checked
even if the node does not have any forward configured. This
lead to a problem as there might be no forwarding delay configured.

Moreover, ensure via DB contraint that forwarding_delay is configured
if a forwarding is enabled for this node.