Add forward on unavailable
Add a new forwarding option ON_UNAVAILABLE that
behaves almost as ON_BUSY but with one difference.

Forward ON_BUSY populates all earlier call legs with
the X-No-Call-Wait option that causes stage2 routing
to reject incoming calls if the line is busy.
Forward ON_UNAVILABLE will not do this, so call waiting
can happen on a called extension (if configured for this
extension). Only if the extension is really unreachable
or cannot be notified of a secondary call because
call waiting is disabled, the forward will take action.