Fix multiring with delays
In case that the first multiring forkrank was equipped with
a delay, this delay was just ignored. Since we do support
multiring participants where also the first group has a delay,
this needed an adaption in the rank generation logik.
If the first forkrank of a multiring participant is now of type
NEXT or DROP, an according separator between the main extension
and the multiring forkranks will be inserted. In case of a DEFAULT
forkrank, everything stays as is.