Add static stage2 routing
Applications need to call special yate targets such as the external/
module. This was only possible via a special entry in yate's
regexroute.conf. This patch adds a user type "static" to the stage2
routing tables with a static routing target. The stage2 routing will
then not do a SIP lookup for the user but rather directly call the
static target.