Stage1: Fix empty outgoing_name
If a faked caller id is set via outgoing_extension in the database entry for
a source extension and no outgoing_name is set, the callername was populated
with None. The yate message serializer doesn't like that. So, we check for
this scenario and keep the original callername in this case.